src/requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
numpy==1.14.3
pandas==0.22.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
